# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User registration API",
        "description": "This is an example API for users registration",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Authorize user",
                "description": "register user",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "JWT authorization token",
                                            "type": "string",
                                            "example": "1|fSPJ2AR0TU0dLB6aiYgtSGHkPnFTfBdh4ltISiSo"
                                        },
                                        "type": {
                                            "description": "Token type",
                                            "type": "string",
                                            "example": "bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Authorize user",
                "description": "Authorizes user by its email and password",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Sanctum authorization token",
                                            "type": "string",
                                            "example": "1|fSPJ2AR0TU0dLB6aiYgtSGHkPnFTfBdh4ltISiSo"
                                        },
                                        "type": {
                                            "description": "Token type",
                                            "type": "string",
                                            "example": "bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RegisterRequest": {
                "properties": {
                    "name": {
                        "description": "User Name",
                        "type": "string",
                        "nullable": false
                    },
                    "email": {
                        "description": "User EMail",
                        "type": "string",
                        "format": "email",
                        "nullable": false
                    },
                    "password": {
                        "description": "User Password",
                        "type": "string",
                        "example": "password",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "properties": {
                    "email": {
                        "description": "User EMail",
                        "type": "string",
                        "format": "email",
                        "nullable": false
                    },
                    "password": {
                        "description": "User Password",
                        "type": "string",
                        "example": "password",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}